smartbill-scanner/requirements.txt
pytesseract
opencv-python
pillow
pandas
scikit-learn
streamlit 